Added class search list to separate the search logic from the UI
